+++ src/main/java/org/kumoricon/model/computer/Computer.java
@@ -9,8 +9,22 @@ public class Computer {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
+    /**
+     * Client computer's IP address
+     */
     private String ipAddress;
+    /**
+     * Name of printer as it is installed on the server
+     */
     private String printerName;
+    /**
+     * Horizontal page offset in points (1/72 inch) when printing pages. >0 moves right, <0 moves left
+     */
+    private Integer xOffset;
+    /**
+     * Vertical page offset in points (1/72 inch) when printing pages. >0 moves up, <0 moves down
+     */
+    private Integer yOffset;
 
     public Integer getId() { return id; }
     public void setId(Integer id) { this.id = id; }
@@ -20,6 +34,12 @@ public class Computer {
 
     public String getPrinterName() { return printerName; }
     public void setPrinterName(String printerName) { this.printerName = printerName; }
+
+    public Integer getxOffset() { return xOffset; }
+    public void setxOffset(Integer xOffset) { this.xOffset = xOffset; }
+
+    public Integer getyOffset() { return yOffset; }
+    public void setyOffset(Integer yOffset) { this.yOffset = yOffset; }
 
     public String toString() {
         return String.format("[Computer %s: %s]", id, ipAddress);